Move test helpers (allowed direct FS ops) to thelper package
