+++ internal/pkg/client/client_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
 
+	"github.com/keboola/keboola-as-code/internal/pkg/thelper"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils"
 )
 
@@ -48,7 +49,7 @@ func TestSimpleRequest(t *testing.T) {
 	assert.NoError(t, response.Err())
 	assert.Equal(t, "test", response.String())
 	expected := "DEBUG  HTTP\tGET https://example.com | 200 | %s"
-	utils.AssertWildcards(t, expected, out.String(), "Unexpected log")
+	thelper.AssertWildcards(t, expected, out.String(), "Unexpected log")
 }
 
 func TestRetry(t *testing.T) {
@@ -70,7 +71,7 @@ func TestRetry(t *testing.T) {
 	assert.Greater(t, c.resty.RetryCount, 2)
 	for i := 1; i <= c.resty.RetryCount; i++ {
 		expected := fmt.Sprintf(`DEBUG  HTTP-ERROR	GET https://example.com | returned http code 504, Attempt %d`, i)
-		assert.Regexp(t, utils.WildcardToRegexp(expected), logs)
+		assert.Regexp(t, thelper.WildcardToRegexp(expected), logs)
 	}
 
 	// Error is logged
@@ -78,7 +79,7 @@ func TestRetry(t *testing.T) {
 		`DEBUG  HTTP-ERROR	GET https://example.com | returned http code 504, Attempt %d`,
 		1+c.resty.RetryCount,
 	)
-	assert.Regexp(t, utils.WildcardToRegexp(expected), logs)
+	assert.Regexp(t, thelper.WildcardToRegexp(expected), logs)
 }
 
 func TestDoNotRetry(t *testing.T) {
@@ -101,7 +102,7 @@ func TestDoNotRetry(t *testing.T) {
 
 	// Error is logged
 	expected := "DEBUG  HTTP-ERROR\tGET https://example.com | returned http code 403, Attempt 1\n"
-	utils.AssertWildcards(t, expected, logs, "Unexpected log")
+	thelper.AssertWildcards(t, expected, logs, "Unexpected log")
 }
 
 func TestVerboseHideSecret(t *testing.T) {
@@ -141,7 +142,7 @@ test
 DEBUG  HTTP	GET https://example.com | 200 | %s
 
 `
-	utils.AssertWildcards(t, expectedLog, out.String(), "Unexpected log")
+	thelper.AssertWildcards(t, expectedLog, out.String(), "Unexpected log")
 }
 
 func getMockedClientAndLogs(t *testing.T, verbose bool) (*Client, *zap.SugaredLogger, *utils.Writer) {
