+++ internal/pkg/remote/test_project.go
@@ -17,6 +17,7 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/filesystem"
 	"github.com/keboola/keboola-as-code/internal/pkg/fixtures"
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
+	"github.com/keboola/keboola-as-code/internal/pkg/thelper"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils"
 )
 
@@ -58,7 +59,7 @@ func newTestProject(t *testing.T, api *StorageApi, stateFilePath string) *testPr
 	}
 
 	// Create project ID
-	if utils.TestProjectId() != api.ProjectId() {
+	if thelper.TestProjectId() != api.ProjectId() {
 		assert.FailNow(t, "TEST_PROJECT_ID and token project id are different.")
 	}
 
@@ -212,7 +213,7 @@ func (p *testProject) setEnv(key string, value string) {
 }
 
 func (p *testProject) logf(format string, a ...interface{}) {
-	if utils.TestIsVerbose() {
+	if thelper.TestIsVerbose() {
 		a = append([]interface{}{p.t.Name()}, a...)
 		p.t.Logf("Fixtures[%s]: "+format, a...)
 	}
