+++ internal/pkg/state/local_test.go
@@ -18,11 +18,12 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/manifest"
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
 	"github.com/keboola/keboola-as-code/internal/pkg/remote"
+	"github.com/keboola/keboola-as-code/internal/pkg/thelper"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils"
 )
 
 func TestLoadLocalStateMinimal(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "minimal")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -44,7 +45,7 @@ func TestLoadLocalStateMinimal(t *testing.T) {
 }
 
 func TestLoadLocalStateComplex(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "complex")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -98,7 +99,7 @@ func TestLoadLocalStateComplex(t *testing.T) {
 }
 
 func TestLoadLocalStateAllowedBranches(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "minimal")
 	m.Content.AllowedBranches = model.AllowedBranches{"main"}
 	state := loadLocalTestState(t, m)
@@ -107,7 +108,7 @@ func TestLoadLocalStateAllowedBranches(t *testing.T) {
 }
 
 func TestLoadLocalStateAllowedBranchesError(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "complex")
 	m.Content.AllowedBranches = model.AllowedBranches{"main"}
 	state := loadLocalTestState(t, m)
@@ -116,7 +117,7 @@ func TestLoadLocalStateAllowedBranchesError(t *testing.T) {
 }
 
 func TestLoadLocalStateBranchMissingMetaJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "branch-missing-meta-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -125,7 +126,7 @@ func TestLoadLocalStateBranchMissingMetaJson(t *testing.T) {
 }
 
 func TestLoadLocalStateBranchMissingDescription(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "branch-missing-description")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -134,7 +135,7 @@ func TestLoadLocalStateBranchMissingDescription(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigMissingConfigJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-missing-config-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -143,7 +144,7 @@ func TestLoadLocalStateConfigMissingConfigJson(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigMissingMetaJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-missing-meta-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -152,7 +153,7 @@ func TestLoadLocalStateConfigMissingMetaJson(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigMissingDescription(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-missing-description")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -161,7 +162,7 @@ func TestLoadLocalStateConfigMissingDescription(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigRowMissingConfigJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-row-missing-config-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -170,7 +171,7 @@ func TestLoadLocalStateConfigRowMissingConfigJson(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigRowMissingMetaJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-row-missing-meta-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -179,7 +180,7 @@ func TestLoadLocalStateConfigRowMissingMetaJson(t *testing.T) {
 }
 
 func TestLoadLocalStateBranchInvalidMetaJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "branch-invalid-meta-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -188,7 +189,7 @@ func TestLoadLocalStateBranchInvalidMetaJson(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigRowMissingDescription(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-row-missing-description")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -197,7 +198,7 @@ func TestLoadLocalStateConfigRowMissingDescription(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigInvalidConfigJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-invalid-config-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -206,7 +207,7 @@ func TestLoadLocalStateConfigInvalidConfigJson(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigInvalidMetaJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-invalid-meta-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -215,7 +216,7 @@ func TestLoadLocalStateConfigInvalidMetaJson(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigRowInvalidConfigJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-row-invalid-config-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -224,7 +225,7 @@ func TestLoadLocalStateConfigRowInvalidConfigJson(t *testing.T) {
 }
 
 func TestLoadLocalStateConfigRowInvalidMetaJson(t *testing.T) {
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 	m := loadManifest(t, "config-row-invalid-meta-json")
 	state := loadLocalTestState(t, m)
 	assert.NotNil(t, state)
@@ -284,7 +285,7 @@ func loadManifest(t *testing.T, projectDirName string) *manifest.Manifest {
 	if err != nil {
 		t.Fatalf("Copy error: %s", err)
 	}
-	utils.ReplaceEnvsDir(projectDir, nil)
+	thelper.ReplaceEnvsDir(projectDir, nil)
 
 	// Load manifest
 	fs, err := aferofs.NewLocalFs(zap.NewNop().Sugar(), projectDir, ".")
