+++ internal/pkg/thelper/wildcards.go
@@ -1,4 +1,4 @@
-package utils
+package thelper
 
 import (
 	"os"
