+++ internal/pkg/thelper/wildcards_test.go
@@ -1,4 +1,4 @@
-package utils
+package thelper
 
 import (
 	"bytes"
