+++ test/functional_test.go
@@ -25,11 +25,12 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/manifest"
 	"github.com/keboola/keboola-as-code/internal/pkg/remote"
 	"github.com/keboola/keboola-as-code/internal/pkg/state"
+	"github.com/keboola/keboola-as-code/internal/pkg/thelper"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils"
 )
 
 // EnvTicketProvider allows you to generate new unique IDs via an ENV variable in the test.
-func CreateEnvTicketProvider(api *remote.StorageApi) utils.EnvProvider {
+func CreateEnvTicketProvider(api *remote.StorageApi) thelper.EnvProvider {
 	return func(name string) string {
 		name = strings.Trim(name, "%")
 		nameRegexp := regexpcache.MustCompile(`^TEST_NEW_TICKET_\d+$`)
@@ -43,7 +44,7 @@ func CreateEnvTicketProvider(api *remote.StorageApi) utils.EnvProvider {
 			return ticket.Id
 		}
 
-		return utils.DefaultEnvProvider(name)
+		return thelper.DefaultEnvProvider(name)
 	}
 }
 
@@ -71,7 +72,7 @@ func TestFunctional(t *testing.T) {
 // RunFunctionalTest runs one functional test.
 func RunFunctionalTest(t *testing.T, testDir, workingDir string, binary string) {
 	t.Helper()
-	defer utils.ResetEnv(t, os.Environ())
+	defer thelper.ResetEnv(t, os.Environ())
 
 	// Clean working dir
 	assert.NoError(t, os.RemoveAll(workingDir))
@@ -80,7 +81,7 @@ func RunFunctionalTest(t *testing.T, testDir, workingDir string, binary string) 
 
 	// Copy all from "in" dir to "runtime" dir
 	inDir := filepath.Join(testDir, "in")
-	if !utils.FileExists(inDir) {
+	if !thelper.FileExists(inDir) {
 		t.Fatalf("Missing directory \"%s\".", inDir)
 	}
 	err := copy.Copy(inDir, workingDir)
@@ -93,7 +94,7 @@ func RunFunctionalTest(t *testing.T, testDir, workingDir string, binary string) 
 
 	// Setup KBC project state
 	projectStateFilePath := filepath.Join(testDir, "initial-state.json")
-	if utils.IsFile(projectStateFilePath) {
+	if thelper.IsFile(projectStateFilePath) {
 		remote.SetStateOfTestProject(t, api, projectStateFilePath)
 	}
 
@@ -101,11 +102,11 @@ func RunFunctionalTest(t *testing.T, testDir, workingDir string, binary string) 
 	envProvider := CreateEnvTicketProvider(api)
 
 	// Replace all %%ENV_VAR%% in all files in the working directory
-	utils.ReplaceEnvsDir(workingDir, envProvider)
+	thelper.ReplaceEnvsDir(workingDir, envProvider)
 
 	// Load command arguments from file
 	argsFile := filepath.Join(testDir, "args")
-	argsStr := utils.ReplaceEnvsString(strings.TrimSpace(utils.GetFileContent(argsFile)), nil)
+	argsStr := thelper.ReplaceEnvsString(strings.TrimSpace(thelper.GetFileContent(argsFile)), nil)
 	args, err := shlex.Split(argsStr)
 	if err != nil {
 		t.Fatalf("Cannot parse args \"%s\": %s", argsStr, err)
@@ -117,7 +118,7 @@ func RunFunctionalTest(t *testing.T, testDir, workingDir string, binary string) 
 	cmd.Dir = workingDir
 	cmd.Stdout = &stdout
 	cmd.Stderr = &stderr
-	if utils.TestIsVerbose() {
+	if thelper.TestIsVerbose() {
 		cmd.Stdout = io.MultiWriter(cmd.Stdout, os.Stdout)
 		cmd.Stderr = io.MultiWriter(cmd.Stderr, os.Stderr)
 	}
@@ -172,11 +173,11 @@ func GetTestDirs(t *testing.T, root string) []string {
 		}
 
 		// Skip hidden
-		if utils.IsIgnoredFile(path, d) {
+		if thelper.IsIgnoredFile(path, d) {
 			return nil
 		}
-		if utils.IsIgnoredDir(path, d) {
-			return fs.SkipDir
+		if thelper.IsIgnoredDir(path, d) {
+			return filepath.SkipDir
 		}
 
 		// Skip sub-directories
@@ -199,7 +200,7 @@ func GetTestDirs(t *testing.T, root string) []string {
 func AssertExpectations(
 	t *testing.T,
 	api *remote.StorageApi,
-	envProvider utils.EnvProvider,
+	envProvider thelper.EnvProvider,
 	testDir string,
 	workingDir string,
 	exitCode int,
@@ -209,9 +210,9 @@ func AssertExpectations(
 	t.Helper()
 
 	// Compare expected values
-	expectedStdout := utils.ReplaceEnvsString(utils.GetFileContent(filepath.Join(testDir, "expected-stdout")), nil)
-	expectedStderr := utils.ReplaceEnvsString(utils.GetFileContent(filepath.Join(testDir, "expected-stderr")), nil)
-	expectedCodeStr := utils.GetFileContent(filepath.Join(testDir, "expected-code"))
+	expectedStdout := thelper.ReplaceEnvsString(thelper.GetFileContent(filepath.Join(testDir, "expected-stdout")), nil)
+	expectedStderr := thelper.ReplaceEnvsString(thelper.GetFileContent(filepath.Join(testDir, "expected-stderr")), nil)
+	expectedCodeStr := thelper.GetFileContent(filepath.Join(testDir, "expected-code"))
 	expectedCode, _ := strconv.ParseInt(strings.TrimSpace(expectedCodeStr), 10, 32)
 	assert.Equal(
 		t,
@@ -223,12 +224,12 @@ func AssertExpectations(
 	)
 
 	// Assert STDOUT and STDERR
-	utils.AssertWildcards(t, expectedStdout, stdout, "Unexpected STDOUT.")
-	utils.AssertWildcards(t, expectedStderr, stderr, "Unexpected STDERR.")
+	thelper.AssertWildcards(t, expectedStdout, stdout, "Unexpected STDOUT.")
+	thelper.AssertWildcards(t, expectedStderr, stderr, "Unexpected STDERR.")
 
 	// Expected state dir
 	expectedDirOrg := filepath.Join(testDir, "out")
-	if !utils.FileExists(expectedDirOrg) {
+	if !thelper.FileExists(expectedDirOrg) {
 		t.Fatalf("Missing directory \"%s\".", expectedDirOrg)
 	}
 
@@ -238,17 +239,17 @@ func AssertExpectations(
 	if err != nil {
 		t.Fatalf("Copy error: %s", err)
 	}
-	utils.ReplaceEnvsDir(expectedDir, envProvider)
+	thelper.ReplaceEnvsDir(expectedDir, envProvider)
 
 	// Compare actual and expected dirs
-	utils.AssertDirectoryContentsSame(t, expectedDir, workingDir)
+	thelper.AssertDirectoryContentsSame(t, expectedDir, workingDir)
 
 	// Check project state
 	expectedStatePath := filepath.Join(testDir, "expected-state.json")
-	if utils.IsFile(expectedStatePath) {
+	if thelper.IsFile(expectedStatePath) {
 		// Read expected state
 		expectedSnapshot := &fixtures.ProjectSnapshot{}
-		err = json.ReadFile(testDir, "expected-state.json", expectedSnapshot, "expected project state")
+		content, err := thelper.ReadFile(testDir, "expected-state.json", "expected project state")
 		if err != nil {
 			assert.FailNow(t, err.Error())
 		}
@@ -276,7 +277,7 @@ func AssertExpectations(
 		}
 
 		// Write actual state
-		err = json.WriteFile(workingDir, "actual-state.json", actualSnapshot, "test project state")
+		err = thelper.WriteFile(workingDir, "actual-state.json", json.MustEncodeString(actualSnapshot, true), "test project state")
 		if err != nil {
 			assert.FailNow(t, err.Error())
 		}
